Extract exposeEvents functions.
