+++ Core/tests/InstructionsParameterMover.cpp
@@ -15,6 +15,7 @@
 #include "GDCore/Events/Builtin/StandardEvent.h"
 #include "GDCore/Extensions/Builtin/SpriteExtension/SpriteObject.h"
 #include "GDCore/Extensions/Platform.h"
+#include "GDCore/IDE/EventsExposer.h"
 #include "GDCore/IDE/Project/ArbitraryResourceWorker.h"
 #include "GDCore/IDE/Project/ProjectResourcesAdder.h"
 #include "GDCore/IDE/WholeProjectRefactorer.h"
@@ -51,7 +52,7 @@ TEST_CASE("InstructionsParameterMover", "[common][events]") {
     gd::InstructionsParameterMover mover(
         project, "MyExtension::DoSomething", 0, 2);
 
-    gd::WholeProjectRefactorer::ExposeProjectEvents(project, mover);
+    gd::EventsExposer::ExposeProjectEvents(project, mover);
     REQUIRE(insertedInstruction.GetParameter(0).GetPlainString() == "Param2");
     REQUIRE(insertedInstruction.GetParameter(1).GetPlainString() == "Param3");
     REQUIRE(insertedInstruction.GetParameter(2).GetPlainString() == "Param1");
@@ -60,7 +61,7 @@ TEST_CASE("InstructionsParameterMover", "[common][events]") {
     gd::InstructionsParameterMover mover(
         project, "MyExtension::DoSomething", 0, 99);
 
-    gd::WholeProjectRefactorer::ExposeProjectEvents(project, mover);
+    gd::EventsExposer::ExposeProjectEvents(project, mover);
     REQUIRE(insertedInstruction.GetParameter(0).GetPlainString() == "Param2");
     REQUIRE(insertedInstruction.GetParameter(1).GetPlainString() == "Param3");
     REQUIRE(insertedInstruction.GetParameter(2).GetPlainString() == "Param1");
@@ -69,7 +70,7 @@ TEST_CASE("InstructionsParameterMover", "[common][events]") {
     gd::InstructionsParameterMover mover(
         project, "MyExtension::DoSomething", 99, 2);
 
-    gd::WholeProjectRefactorer::ExposeProjectEvents(project, mover);
+    gd::EventsExposer::ExposeProjectEvents(project, mover);
     REQUIRE(insertedInstruction.GetParameter(0).GetPlainString() == "Param1");
     REQUIRE(insertedInstruction.GetParameter(1).GetPlainString() == "Param2");
     REQUIRE(insertedInstruction.GetParameter(2).GetPlainString() == "Param3");
