+++ GDevelop.js/Bindings/Wrapper.cpp
@@ -51,6 +51,7 @@
 #include <GDCore/IDE/Project/ResourcesMergingHelper.h>
 #include <GDCore/IDE/Project/ResourcesRenamer.h>
 #include <GDCore/IDE/PropertyFunctionGenerator.h>
+#include <GDCore/IDE/EventsExposer.h>
 #include <GDCore/IDE/WholeProjectRefactorer.h>
 #include <GDCore/IDE/UnfilledRequiredBehaviorPropertyProblem.h>
 #include <GDCore/Project/Behavior.h>
