+++ GDevelop.js/types/gdeventsexposer.js
@@ -1,0 +1,6 @@
+// Automatically generated by GDevelop.js/scripts/generate-types.js
+declare class gdEventsExposer {
+  static exposeProjectEvents(project: gdProject, worker: gdArbitraryEventsWorker): void;
+  delete(): void;
+  ptr: number;
+};
