+++ GDevelop.js/types/gdwholeprojectrefactorer.js
@@ -1,6 +1,5 @@
 // Automatically generated by GDevelop.js/scripts/generate-types.js
 declare class gdWholeProjectRefactorer {
-  static exposeProjectEvents(project: gdProject, worker: gdArbitraryEventsWorker): void;
   static renameEventsFunctionsExtension(project: gdProject, eventsFunctionsExtension: gdEventsFunctionsExtension, oldName: string, newName: string): void;
   static renameEventsFunction(project: gdProject, eventsFunctionsExtension: gdEventsFunctionsExtension, oldName: string, newName: string): void;
   static renameBehaviorEventsFunction(project: gdProject, eventsFunctionsExtension: gdEventsFunctionsExtension, eventsBasedBehavior: gdEventsBasedBehavior, oldName: string, newName: string): void;
