+++ newIDE/app/src/Leaderboard/useLeaderboardReplacer.js
@@ -315,7 +315,7 @@ export const useLeaderboardReplacer = (): UseLeaderboardReplacerOutput => {
         );
         renamedLeaderboardsMap.delete();
 
-        gd.WholeProjectRefactorer.exposeProjectEvents(
+        gd.EventsExposer.exposeProjectEvents(
           project,
           eventsLeaderboardReplacer
         );
@@ -371,7 +371,7 @@ export const useLeaderboardReplacer = (): UseLeaderboardReplacerOutput => {
       setGameId(sourceGameId);
 
       const leaderboardsLister = new gd.EventsLeaderboardsLister(project);
-      gd.WholeProjectRefactorer.exposeProjectEvents(
+      gd.EventsExposer.exposeProjectEvents(
         project,
         leaderboardsLister
       );
